Refactored file structure + navbar
